Fixing delete

+++ app/static/feed-script.js
@@ -15,19 +15,22 @@ function tweetLimitCounter(){
 
 
 function deleteTwote(this_ele) {
-  console.log("delete clicked");
-  const ele = {
-    name: document.getElementById("twotecontent").id,
-  };
-  twote_id = this_ele.id;
-  fetch(`http://127.0.0.1:5000/twote?twote_id=${twote_id}`, {
-    method: "DELETE",
-    headers: { "Content-Type": "application/json" },
-    body: JSON.stringify(ele),
-  })
-    .then((response) => response.json())
-    .then((data) => console.log(data))
-    .then(() => location.reload());
+  if(confirm("Are you sure you want to delete this twote?")){
+    console.log("delete clicked");
+    const ele = {
+      name: document.getElementById("twotecontent").id,
+    };
+    twote_id = this_ele.id;
+    fetch(`http://127.0.0.1:5000/twote?twote_id=${twote_id}`, {
+      method: "DELETE",
+      headers: { "Content-Type": "application/json",
+                  "Accept": "application/json"},
+      body: JSON.stringify(ele),
+    })
+      //.then((response) => response.json())
+      //.then((data) => console.log(data))
+      .then(() =>{window.location.reload()});
+  }
 }
 
 function cannotOpen(){
